requirements.txt
yfinance
pandas
pandas-ta
python-telegram-bot
python-dotenv
